% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_passing.R
\name{calc_down_messages}
\alias{calc_down_messages}
\title{calculate the down messages}
\usage{
calc_down_messages(
  phy,
  time_scale,
  argument,
  E_list,
  up_messages,
  T_vector,
  non_negativity_cutoff
)
}
\arguments{
\item{phy}{a phylo object (tree object returned by nj)}

\item{time_scale}{time normalization factor}

\item{argument}{a list of birth, death and mutation data}

\item{E_list}{pre-calculated E_list (see the next step)}

\item{up_messages}{pre-calculated up messages}
}
\value{
matrix of "messages" sent to a node by its ancestral and sibling lineages, rows: node; columns: fitness type
}
\description{
calculate the down messages
}
