% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_passing.R
\name{calc_marginal_probabilities}
\alias{calc_marginal_probabilities}
\title{calculate the marginal distribution for each node}
\usage{
calc_marginal_probabilities(
  phy,
  time_scale,
  up_messages,
  down_messages,
  argument
)
}
\arguments{
\item{phy}{a phylo object (tree object returned by nj)}

\item{time_scale}{time normalization factor}

\item{up_messages}{pre-calculated "up-messages" to every nodes}

\item{down_messages}{pre-calculated "down-messages" to every nodes}

\item{argument}{a list of birth, death and mutation data}
}
\value{
a matrix of marginal probablity distribution on fitness type (column) for each node (row)
}
\description{
calculate the marginal distribution for each node
}
