% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_passing.R
\name{calc_up_messages}
\alias{calc_up_messages}
\title{calculate the up messages along each branch to the up node
branch index * up node state}
\usage{
calc_up_messages(
  phy,
  time_scale,
  argument,
  rho,
  d_t,
  E_list,
  T_vector,
  non_negativity_cutoff,
  n_threads
)
}
\arguments{
\item{phy}{a phylo object (tree object returned by nj)}

\item{time_scale}{time normalization factor}

\item{argument}{a list of birth, death and mutation data}

\item{rho}{sampling probability}

\item{d_t}{step size for integration}

\item{E_list}{pre-calculated E_list (see the next step)}

\item{n_threads}{number of threads for parallele computing}
}
\value{
matrix of "messages" sent to a node by its progeny lineages, rows: node; columns: fitness type
}
\description{
calculate the up messages along each branch to the up node
branch index * up node state
}
