% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_density.R
\name{derivative_D}
\alias{derivative_D}
\title{calculate the derivative of D}
\usage{
derivative_D(phi, eta, argument, ini_rank)
}
\arguments{
\item{phi}{value of the function D}

\item{eta}{pre-calculated Ei}

\item{argument}{a list of birth, death and mutation data}

\item{ini_rank}{only keeps the possible transition between states (the density from higher rank to lower rank is zero)}
}
\value{
vector of calculated derivatives of D
}
\description{
calculate the derivative of D
}
