% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_density.R
\name{derivative_E}
\alias{derivative_E}
\title{calculate the derivative of Extinction density}
\usage{
derivative_E(phi, argument)
}
\arguments{
\item{phi}{value of the function E}

\item{argument}{a list of birth, death and mutation data}
}
\value{
vector of calculated derivatives of E(t)
}
\description{
calculate the derivative of Extinction density
}
