% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_inference.R
\name{distanceMatrix2Tree}
\alias{distanceMatrix2Tree}
\title{prepare input of tree from distance matrix
distance matrix input, prepare phy
the distanceMatrix input (tab delimited) currently has this format
cell1    cell2   cell3   cell4  cell5  ...
cell1          0        123     11       450     21    ...
cell2          123      0       34       34      45    ...
...
the last cell represents the normal cell}
\usage{
distanceMatrix2Tree(distanceMatrix)
}
\arguments{
\item{distanceMatrix}{the mutationMatrix file input (comma delimited)}
}
\value{
a phylo object of the tree, with the root edge being removed
}
\description{
prepare input of tree from distance matrix
distance matrix input, prepare phy
the distanceMatrix input (tab delimited) currently has this format
cell1    cell2   cell3   cell4  cell5  ...
cell1          0        123     11       450     21    ...
cell2          123      0       34       34      45    ...
...
the last cell represents the normal cell
}
