% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_density.R
\name{get_D_list}
\alias{get_D_list}
\title{prepare the functions of D(t) for each type transition starting from a given time t_1}
\usage{
get_D_list(rho, T_vector, d_t, argument, E_list, t_1, non_negativity_cutoff)
}
\arguments{
\item{rho}{sampling probability}

\item{T_vector}{time discretization}

\item{d_t}{step size for integration}

\item{argument}{a list of birth, death and mutation data}

\item{E_list}{precalculated E_list (see the next step)}

\item{t_1}{time of the down node (starting node)}
}
\value{
Interpolated functions of D(t) encapsulated in a two dimensional list. the first dimension being the down node (child node) and the second being the upper node (ancestral node).
}
\description{
prepare the functions of D(t) for each type transition starting from a given time t_1
}
