% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{get_full_pars}
\alias{get_full_pars}
\title{Get a full set of LeafRank configuration.}
\usage{
get_full_pars(
  b_rates = NULL,
  d_rates = NULL,
  nu = NULL,
  rho = NULL,
  tau = NULL,
  init = NULL,
  tree,
  model = "default"
)
}
\arguments{
\item{b_rates}{1 x V vector for birth rates at each fitness phenotype}

\item{d_rates}{1 x V vector for death rates at each fitness phenotype}

\item{nu}{coefficient for driver mutation rates}

\item{rho}{sampling probability}

\item{tau}{time scale}

\item{init}{initial condition for original cell}
}
\value{
list of configuration parameters with 1. birth rates, 2. death rates, 3. nu, 4. rho, 5. Time scale
}
\description{
Get a full set of LeafRank configuration.
}
