% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{get_siblings}
\alias{get_siblings}
\title{Helper function to get sibling node}
\usage{
get_siblings(node, tree)
}
\arguments{
\item{node}{current node}

\item{phy}{a phylo object}
}
\value{
node attached to the sibling branch
}
\description{
Helper function to get sibling node
}
