% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{get_ultrametric}
\alias{get_ultrametric}
\title{Access matlab results and generate an ultrametric tree under the WGD assumption.}
\usage{
get_ultrametric(tree, MAT_path, phy_path = NULL)
}
\arguments{
\item{tree}{original distance-matrix based tree.}

\item{MAT_path}{matlab data ('.csv' file) saving path}

\item{phy_path}{tree saving path}
}
\value{
WGD ultrametric tree
}
\description{
Access matlab results and generate an ultrametric tree under the WGD assumption.
}
