% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{get_ultrametric_prepared}
\alias{get_ultrametric_prepared}
\title{Prepared for reconstructing ultrametric tree.}
\usage{
get_ultrametric_prepared(phy, tips_WGD, MAT_path = NULL)
}
\arguments{
\item{phy}{a phylo object (tree object returned by nj or MEDICC2)}

\item{tips_WGD}{a vector records the WGD status for all tips}

\item{MAT_path}{matlab data saving path}
}
\value{
.mat data storing necessary information about a distance-matrix based tree
}
\description{
Prepared for reconstructing ultrametric tree.
}
