% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_density.R
\name{integrate_D}
\alias{integrate_D}
\title{integrate_D is a function to solve differential equations using Runge-Kutta methods
D(t) can be viewed as the density that a node at state i leads to the sampled tree
similar as the "branch propagator" defined in Neher's paper (given an ancestor with fitness y at time t,
the prob that the child is with fitness x and time t')}
\usage{
integrate_D(
  f_initial,
  T_vector,
  argument,
  E_list,
  ini_rank,
  t_1,
  d_t,
  non_negativity_cutoff
)
}
\arguments{
\item{f_initial}{vector of initial values (for each fitness type)}

\item{T_vector}{time discretization}

\item{argument}{a list of birth, death and mutation data}

\item{E_list}{precalculated E(t), acorss a list of different types}

\item{ini_rank}{only keeps the possible transition between state}

\item{t_1}{time of the node closer to the leaves ("down node"), this is to offset E(t) which is dependent on time}

\item{d_t}{step size for integration}
}
\value{
numerically calculated D(t) over the time vector, a matrix with rows representing time and columns representing fitness types.
}
\description{
integrate_D is a function to solve differential equations using Runge-Kutta methods
D(t) can be viewed as the density that a node at state i leads to the sampled tree
similar as the "branch propagator" defined in Neher's paper (given an ancestor with fitness y at time t,
the prob that the child is with fitness x and time t')
}
