% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_density.R
\name{integrate_E}
\alias{integrate_E}
\title{Calculate Extinction density over time
note that E(t) is independent of tree structure
solve differential equations using Runge-Kutta methods (https://en.wikipedia.org/wiki/Runge\%E2\%80\%93Kutta_methods)}
\usage{
integrate_E(f_initial, T_vector, argument, d_t, non_negativity_cutoff)
}
\arguments{
\item{f_initial}{vector of initial values (for each type of cells)}

\item{T_vector}{time discretization}

\item{argument}{a list of birth, death and mutation data (b <- argument[\link{1}]; d <- argument[\link{2}]; mu <- argument[\link{3}])}

\item{d_t}{step size for integration}
}
\value{
numerically calculated E(t) over the time vector, a matrix with rows representing time and columns representing fitness types.
}
\description{
Calculate Extinction density over time
note that E(t) is independent of tree structure
solve differential equations using Runge-Kutta methods (https://en.wikipedia.org/wiki/Runge\%E2\%80\%93Kutta_methods)
}
