% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_density.R
\name{integrate_phi_D}
\alias{integrate_phi_D}
\title{call the function integrate_D to integrate D rootward staring at time t_1}
\usage{
integrate_phi_D(
  rho,
  T_vector,
  argument,
  E_list,
  ini_rank,
  f_initial,
  t_1,
  d_t,
  non_negativity_cutoff
)
}
\arguments{
\item{rho}{sampling probability}

\item{T_vector}{time discretization}

\item{argument}{a list of birth, death and mutation data}

\item{E_list}{precalculated E_list (see the next step)}

\item{ini_rank}{only keeps the possible transition between state}

\item{f_initial}{vector of initial values (for each type of cells)}

\item{t_1}{time of the down node (starting node)}

\item{d_t}{step size for integration}

\item{non_negativity_cutoff}{Cut off threshold}
}
\value{
numerically calculated solution of D(t)
}
\description{
call the function integrate_D to integrate D rootward staring at time t_1
}
