% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_density.R
\name{integrate_phi_E}
\alias{integrate_phi_E}
\title{call the function integrate_E to integrate E(t)}
\usage{
integrate_phi_E(rho, T_vector, argument, d_t, non_negativity_cutoff)
}
\arguments{
\item{rho}{sampling probability}

\item{T_vector}{time discretization}

\item{argument}{a list of birth, death and mutation data}

\item{d_t}{step size for integration}

\item{non_negativity_cutoff}{Cut off threshold}
}
\value{
numerically calculated solution of E(t)
}
\description{
call the function integrate_E to integrate E(t)
}
