% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_density.R
\name{integrate_prop}
\alias{integrate_prop}
\title{calculate the propagator for a branch, a propagator is the probability density from an ancestral node to a child}
\usage{
integrate_prop(
  rho,
  argument,
  t,
  t_1,
  E_list,
  T_vector,
  d_t,
  non_negativity_cutoff
)
}
\arguments{
\item{rho}{sampling probability}

\item{argument}{a list of birth, death and mutation data}

\item{t}{time between up node and down node (length of the branch)}

\item{t_1}{starting time from the child}

\item{E_list}{pre-calculated E_list (see the next step)}

\item{T_vector}{time discretization}

\item{d_t}{step size for integration}

\item{non_negativity_cutoff}{Cut off threshold}
}
\value{
a matrix of integrated D values of a branch (ancestral type to the child type)
}
\description{
calculate the propagator for a branch, a propagator is the probability density from an ancestral node to a child
}
