% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_inference.R
\name{ith.Fitness}
\alias{ith.Fitness}
\title{pipeline to do fitness inference}
\usage{
ith.Fitness(
  phy,
  outFile,
  rho,
  d_t,
  time_scale,
  b_rate,
  d_rate,
  mu,
  T_vector,
  non_negativity_cutoff
)
}
\arguments{
\item{phy}{a phylo object (tree object returned by nj)}

\item{outFile}{output File with full path}

\item{rho}{sampling probability}

\item{d_t}{step size for integration}

\item{time_scale}{time normalization factor}

\item{b_rate}{birth rates for types (vector)}

\item{d_rate}{death rates for types (vector)}

\item{mu}{driver mutation rates for types (vector)}

\item{T_vector}{time discretization (vector)}

\item{non_negativity_cutoff}{}
}
\value{
the mean fitness vector
}
\description{
pipeline to do fitness inference
}
