% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{label_internal_nodes}
\alias{label_internal_nodes}
\title{Label the WGD status for internal nodes and branches based on the tip annotations.
This function is based on the `anc assumption', in which a branch is identified as WGD if its descendant node is WGD.}
\usage{
label_internal_nodes(tree)
}
\arguments{
\item{tree}{a phylo object with feature node_wgd, indicating the WGD status of tip nodes.}
}
\value{
a phylo object with internal nodes and branches labeled accordingly for WGD status.
}
\description{
Label the WGD status for internal nodes and branches based on the tip annotations.
This function is based on the `anc assumption', in which a branch is identified as WGD if its descendant node is WGD.
}
