% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_passing.R
\name{log_normalize}
\alias{log_normalize}
\title{convert numbers in log form to a probability measure}
\usage{
log_normalize(prob_vector)
}
\arguments{
\item{prob_vector}{a vector of probablity-related values that need to be normalized}
}
\value{
a vector of normalized probablities
}
\description{
convert numbers in log form to a probability measure
}
