% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_passing.R
\name{mean_fitness}
\alias{mean_fitness}
\title{calculate the marginal distribution for each node}
\usage{
mean_fitness(phy, marginal_prob, argument)
}
\arguments{
\item{phy}{a phylo object (tree object returned by nj)}

\item{marginal_prob}{a matrix of marginal probability distribution on fitness type (column) for each node (row)}

\item{argument}{a list of birth, death and mutation data}
}
\value{
vector of mean fitness of each node
}
\description{
calculate the marginal distribution for each node
}
