% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{node_num}
\alias{node_num}
\title{return the number of nodes (internal and external)}
\usage{
node_num(phy)
}
\arguments{
\item{phy}{a phylo object (tree object returned by nj)}
}
\value{
the number of nodes in the tree
}
\description{
return the number of nodes (internal and external)
}
