% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{node_status}
\alias{node_status}
\title{Check if a node is internal or external}
\usage{
node_status(phy, node_id)
}
\arguments{
\item{phy}{a phylo object (tree object returned by nj)}

\item{node_id}{the id of the node of interest}
}
\value{
(logical) TRUE: internal, FALSE: external
}
\description{
Check if a node is internal or external
}
