% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{node_time_to_present}
\alias{node_time_to_present}
\title{transform the tree into an ultrametric tree
re-scale the branch length (measured by mutations) to the real time scale}
\usage{
node_time_to_present(phy, time_scale)
}
\arguments{
\item{phy}{a phylo object (tree object returned by nj)}

\item{time_scale}{time normalization factor}
}
\value{
vector of normalized branch lengths in real time scale
}
\description{
time = 0 for leaves
}
