% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_inference.R
\name{rank_diff}
\alias{rank_diff}
\title{Compare the ranking inferred by two different sets of parameters}
\usage{
rank_diff(result_1, result_2)
}
\arguments{
\item{result_1}{mean fitness inferred by the first set of parameters}

\item{result_2}{mean fitness inferred by the second set of parameters}
}
\value{
a Kendall-tau like distance
}
\description{
Compare the ranking inferred by two different sets of parameters
}
