% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_passing.R
\name{up_m_des}
\alias{up_m_des}
\title{obtain the message upcoming from the down_node by taking into account its branching
each element represents a state of down node of the branch of interest
handle the branching from the child}
\usage{
up_m_des(up_messages, down_node, phy, argument)
}
\arguments{
\item{up_messages}{pre-calculated up messages}

\item{down_node}{index of the down node}

\item{phy}{a phylo object (tree object returned by nj)}

\item{argument}{a list of birth, death and mutation data}
}
\value{
the message (across fitness states) upcoming from the down node
}
\description{
obtain the message upcoming from the down_node by taking into account its branching
each element represents a state of down node of the branch of interest
handle the branching from the child
}
