% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_passing.R
\name{up_m_sib}
\alias{up_m_sib}
\title{obtain the message upcoming from the up_node by taking into account the branching of its parental node
for calculating down message, we need to know the branching staring the mother node (sibling lineages)
convert numbers in log form to a probability measure}
\usage{
up_m_sib(up_messages, down_messages, up_node, phy, argument, i)
}
\arguments{
\item{up_messages}{pre-calculated up messages}

\item{down_messages}{pre-calculated down messages}

\item{up_node}{index of the up node}

\item{phy}{a phylo object (tree object returned by nj)}

\item{argument}{a list of birth, death and driver aberration data}

\item{i}{branch index}
}
\value{
the message (across fitness states) coming from the up_node
}
\description{
obtain the message upcoming from the up_node by taking into account the branching of its parental node
for calculating down message, we need to know the branching staring the mother node (sibling lineages)
convert numbers in log form to a probability measure
}
