% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{get_depth}
\alias{get_depth}
\title{Return the depth of each nodes (depth measures the distance between the root and a node)}
\usage{
get_depth(phy)
}
\arguments{
\item{phy}{a phylo object (tree object returned by nj)}
}
\value{
vector of depths of nodes
}
\description{
Return the depth of each nodes (depth measures the distance between the root and a node)
}
