% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{get_descendant}
\alias{get_descendant}
\title{get the ids of descendants of an internal node
enter an external node can lead to infinite loop (modify later)}
\usage{
get_descendant(phy, node_id)
}
\arguments{
\item{phy}{a phylo object (tree object returned by nj)}

\item{node_id}{the id of the node of interest}
}
\value{
vector ids of the two children nodes
}
\description{
get the ids of descendants of an internal node
enter an external node can lead to infinite loop (modify later)
}
