% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{get_sibling}
\alias{get_sibling}
\title{get the sibling branch of a branch
need to optimize}
\usage{
get_sibling(phy)
}
\arguments{
\item{phy}{a phylo object (tree object returned by nj)}
}
\value{
vector of branch ids corresponding to the sibling branch (of ordered branch ids)
}
\description{
get the sibling branch of a branch
need to optimize
}
