% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_inference.R
\name{mutationMatrix2Tree}
\alias{mutationMatrix2Tree}
\title{calculate the up messages along each branch to the up node
branch index * up node state
mutation input, prepare phy
the treeFile input (tab delimited) currently has this format
mutationID   cell1    cell2   cell3   cell4  cell5  ...
12345          1        1       0       0      0    ...
67890          0        1       1       1      0    ...
...}
\usage{
mutationMatrix2Tree(treeFile)
}
\arguments{
\item{treeFile}{the mutationMatrix file input (tab delimited)}
}
\value{
a phylo object of the tree, with the root edge being removed
}
\description{
calculate the up messages along each branch to the up node
branch index * up node state
mutation input, prepare phy
the treeFile input (tab delimited) currently has this format
mutationID   cell1    cell2   cell3   cell4  cell5  ...
12345          1        1       0       0      0    ...
67890          0        1       1       1      0    ...
...
}
